pyproject.toml
[tool.poetry]
name = "crewai-sheets-ui"
version = "0.5.3"
description = "Use Google sheets as UI for CrewAI"
authors = ["yuriwa"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12.5"
crewai = "*"
crewai-tools = "*"
pandas = "*"
ollama = "*"
langchain-anthropic = "*"
langchain-groq = "<=0.1.3"
rich = "*"
restrictedpython = "*"
langchain-community = "*"
duckduckgo-search = "*"
wikipedia = "*"
sentry-sdk = "*"
open-interpreter = "*"
pre-commit = "^3.8.0"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.1"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
